chore: refactor so that connector controller handles selection instead
