+++ packages/appkit-utils/exports/index.ts
@@ -7,7 +7,6 @@ export { PresetsUtil } from '../src/PresetsUtil.js'
 export { HelpersUtil } from '../src/HelpersUtil.js'
 export { ErrorUtil } from '../src/ErrorUtil.js'
 export { LoggerUtil } from '../src/LoggerUtil.js'
-export { MobileWalletUtil } from '../src/MobileWallet.js'
 export { CaipNetworksUtil, getBlockchainApiRpcUrl } from '../src/CaipNetworkUtil.js'
 export type { SocialProvider } from '../src/TypeUtil.js'
