+++ packages/scaffold-ui/src/partials/w3m-connect-featured-widget/index.ts
@@ -3,9 +3,8 @@ import { property } from 'lit/decorators.js'
 import { ifDefined } from 'lit/directives/if-defined.js'
 
 import type { WcWallet } from '@reown/appkit-core'
-import { ApiController, AssetUtil } from '@reown/appkit-core'
+import { ApiController, AssetUtil, ConnectorController } from '@reown/appkit-core'
 import { customElement } from '@reown/appkit-ui'
-import { MobileWalletUtil } from '@reown/appkit-utils'
 
 import { WalletUtil } from '../../utils/WalletUtil.js'
 
@@ -52,7 +51,7 @@ export class W3mConnectFeaturedWidget extends LitElement {
 
   // -- Private Methods ----------------------------------- //
   private onConnectWallet(wallet: WcWallet) {
-    MobileWalletUtil.handleMobileWalletRedirection(wallet)
+    ConnectorController.selectWalletConnector(wallet)
   }
 }
